cleanup: Refactor into modules This change moves models/renshuu_api/windows code into separate file.
